Create TowerOfHanoi.java
This program lists the steps to solve the Tower of Hanoi problem with 3 pegs. The number of disks to start with can be input as an argument. The steps are printed on screen up to 10 disks. For more than that, the steps
are written to a file as the output grows exponentially with the number of disks.

The solution to Tower of Hanoi takes exponential time. Big O(n) = 2 ^ n